+++ tests/unit/Factory/Helper/FactoryBase.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Phalcon\Test\Unit\Factory\Helper;
+
+use Phalcon\Config\Adapter\Ini;
+use Phalcon\Test\Module\UnitTest;
+
+/**
+ * Class FactoryBase
+ * @package Phalcon\Test\unit\Factory\Helper
+ */
+abstract class FactoryBase extends UnitTest
+{
+    /**
+     * @var Ini
+     */
+    protected $config;
+    /**
+     * @var array
+     */
+    protected $arrayConfig;
+
+    public function _before()
+    {
+        parent::_before();
+
+        $this->config = new Ini(PATH_DATA."config/factory.ini", INI_SCANNER_NORMAL);
+        $this->arrayConfig = $this->config->toArray();
+    }
+}